Move ImporXML methods to ObserverManagerMenuFile
